+++ tests/orcid/test_extract.py
@@ -1,13 +1,22 @@
 import pytest
+from requests import HTTPError
 
 from mex.common.exceptions import EmptySearchResultError, FoundMoreThanOneError
 from mex.common.orcid.extract import (
     get_data_by_id,
     get_data_by_name,
-    get_orcid_person_by_id,
-    get_orcid_person_by_name,
+    get_orcid_record_by_id,
+    get_orcid_record_by_name,
 )
-from mex.common.orcid.models.person import FamilyName, GivenNames, OrcidPerson
+from mex.common.orcid.models.person import (
+    OrcidEmails,
+    OrcidFamilyName,
+    OrcidGivenNames,
+    OrcidIdentifier,
+    OrcidName,
+    OrcidPerson,
+    OrcidRecord,
+)
 from tests.orcid.conftest import DUMMYDATA_JOHN, DUMMYDATA_VYVY
 
 
@@ -25,18 +34,12 @@ def test_get_data_by_name():
         (
             "Mustermann",
             "Max",
-            pytest.raises(
-                FoundMoreThanOneError,
-                match="Found multiple AD persons for filters {'given-names': 'Max', 'family-name': 'Mustermann'}",
-            ),
+            pytest.raises(FoundMoreThanOneError),
         ),
         (
             "Defgh",
             "Abc",
-            pytest.raises(
-                EmptySearchResultError,
-                match="Cannot find orcid person for filters {'given-names': 'Abc', 'family-name': 'Defgh'}",
-            ),
+            pytest.raises(EmptySearchResultError),
         ),
     ],
     ids=["multiple results", "empty result"],
@@ -47,65 +50,106 @@ def test_search_errors(family_name, given_names, error):
 
 
 @pytest.mark.parametrize(
-    ("string_id", "expected"),
+    ("string_id", "expected", "error"),
     [
-        (
-            "0009-0004-3041-5706",
-            DUMMYDATA_JOHN,
-        ),
-        ("0009-0004-3041-576", {"result": None, "num-found": 0}),
-        ("invalid-orcid-id", {"result": None, "num-found": 0}),
+        ("0009-0004-3041-5706", DUMMYDATA_JOHN, None),
+        ("0009-0004-3041-576", None, pytest.raises(HTTPError)),
+        ("invalid-orcid-id", None, pytest.raises(HTTPError)),
     ],
     ids=["existing person", "non-existing person", "invalid characters"],
 )
-def test_get_data_by_orcid_id(string_id, expected) -> None:
-    result = get_data_by_id(orcid_id=string_id)
-    if result.get("num-found"):
-        assert result.get("num-found") == expected.get("num-found")
-    assert result == expected
+def test_get_data_by_orcid_id(string_id, expected, error) -> None:
+    if error:
+        with error:
+            get_data_by_id(orcid_id=string_id)
+    else:
+        result = get_data_by_id(orcid_id=string_id)
+        assert result == expected
 
 
 @pytest.mark.parametrize(
-    ("given_names", "family_name", "expected_result"),
+    ("given_names", "family_name", "expected_result", "error"),
     [
         (
             "Vyvy",
             "Tran Ngoc",
-            OrcidPerson(
-                orcid_identifier="0009-0004-3041-5706",
-                email=[],
-                given_names=GivenNames(given_names="VyVy", visibility="public"),
-                family_name=FamilyName(family_name="Tran Ngoc", visibility="public"),
+            OrcidRecord(
+                orcid_identifier=OrcidIdentifier(
+                    path="0009-0004-3041-5706",
+                    uri="https://orcid.org/0009-0004-3041-5706",
+                ),
+                person=OrcidPerson(
+                    emails=OrcidEmails(email=[]),
+                    name=OrcidName(
+                        family_name=OrcidFamilyName(value="Tran Ngoc"),
+                        given_names=OrcidGivenNames(value="VyVy"),
+                        visibility="public",
+                    ),
+                ),
             ),
-        )
-    ],
-)
-def test_get_orcid_person_by_name(
-    given_names,
-    family_name,
-    expected_result,
-):
-    result = get_orcid_person_by_name(given_names, family_name)
-    assert result == expected_result
-
-
-@pytest.mark.parametrize(
-    ("orcid_id", "expected_result"),
-    [
+            None,
+        ),
         (
-            "0000-0002-1825-0097",
-            OrcidPerson(
-                orcid_identifier="0000-0002-1825-0097",
-                email=[],
-                given_names=GivenNames(given_names="Josiah", visibility="public"),
-                family_name=FamilyName(family_name="Carberry", visibility="public"),
+            "Abc",
+            "Defgh",
+            None,
+            pytest.raises(
+                EmptySearchResultError,
+            ),
+        ),
+        (
+            "John",
+            "Doe",
+            None,
+            pytest.raises(
+                FoundMoreThanOneError,
             ),
         ),
     ],
+    ids=[],
 )
-def test_get_orcid_person_by_id(
-    orcid_id,
-    expected_result,
-):
-    result = get_orcid_person_by_id(orcid_id)
-    assert result == expected_result
+def test_get_orcid_record_by_name(given_names, family_name, expected_result, error):
+    if error:
+        with error:
+            get_orcid_record_by_name(given_names, family_name)
+    else:
+        result = get_orcid_record_by_name(given_names, family_name)
+        assert result == expected_result
+
+
+@pytest.mark.parametrize(
+    ("orcid_id", "expected_result", "error"),
+    [
+        (
+            "0000-0002-1825-0097",
+            OrcidRecord(
+                orcid_identifier=OrcidIdentifier(
+                    path="0000-0002-1825-0097",
+                    uri="https://orcid.org/0000-0002-1825-0097",
+                ),
+                person=OrcidPerson(
+                    emails=OrcidEmails(email=[]),
+                    name=OrcidName(
+                        family_name=OrcidFamilyName(value="Carberry"),
+                        given_names=OrcidGivenNames(value="Josiah"),
+                        visibility="public",
+                    ),
+                ),
+            ),
+            None,
+        ),
+        (
+            "0000-0001-2345-6789",
+            None,
+            pytest.raises(HTTPError),
+        ),
+    ],
+    ids=["existing id", "non-existing id"],
+)
+def test_get_orcid_record_by_id(orcid_id, expected_result, error):
+    if error:
+        with error:
+            get_orcid_record_by_id(orcid_id)
+    else:
+        result = get_orcid_record_by_id(orcid_id)
+        assert result == expected_result